Move phase loader table from Config to Phase, close #213
