+++ Ash2/Ash2.vcxproj.filters
@@ -264,6 +264,9 @@
     <ClCompile Include="Phase\PhaseStack.cpp">
       <Filter>Source Files\Phase</Filter>
     </ClCompile>
+    <ClCompile Include="Phase\PhaseLoaders.cpp">
+      <Filter>Source Files\Phase</Filter>
+    </ClCompile>
     <ClCompile Include="stdafx.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -1068,6 +1071,9 @@
       <Filter>Header Files\Phase</Filter>
     </ClInclude>
     <ClInclude Include="Phase\IPhase.hpp">
+      <Filter>Header Files\Phase</Filter>
+    </ClInclude>
+    <ClInclude Include="Phase\PhaseLoaders.hpp">
       <Filter>Header Files\Phase</Filter>
     </ClInclude>
     <ClInclude Include="Phase\PhaseStack.hpp">
