+++ Ash2/src/Config/ScenarioData.hpp
@@ -1,6 +1,8 @@
 #pragma once
 #include <Siv3D.hpp>
 
+#include <expected>
+#include <functional>
 #include <memory>
 #include <variant>
 
@@ -15,6 +17,13 @@ struct IPhaseMaker {
   /// @brief 保持しているパラメータから IPhase インスタンスを生成する
   [[nodiscard]] virtual std::unique_ptr<IPhase> make() const = 0;
 };
+
+/// @brief TOML 値から IPhaseMaker を生成する関数の型
+using PhaseLoader =
+    std::function<std::expected<IPhaseMaker::Ptr, String>(const TOMLValue&)>;
+
+/// @brief フェーズ名 → PhaseLoader のマップ
+using PhaseLoaderTable = HashTable<String, PhaseLoader>;
 
 /// @brief シナリオのステップ：フェーズをスタックに積む
 struct StepPush {
@@ -35,5 +44,7 @@ struct ScenarioData {
   HashTable<String, Array<ScenarioStep>> sections;
 
   /// @brief TOML からシナリオデータを生成する
-  [[nodiscard]] static ScenarioData FromToml(const TOMLValue& toml);
+  /// @param loaders フェーズ名 → PhaseLoader のマップ（呼び出し元が用意する）
+  [[nodiscard]] static ScenarioData FromToml(const TOMLValue& toml,
+                                             const PhaseLoaderTable& loaders);
 };
