+++ Ash2/tests/TestScenarioData.cpp
@@ -2,6 +2,7 @@
 #include <ThirdParty/Catch2/catch.hpp>
 
 #include "Config/ScenarioData.hpp"
+#include "Phase/PhaseLoaders.hpp"
 #include "Phase/WaitPhase.hpp"
 
 TEST_CASE("ScenarioData::FromToml - push action creates StepPush") {
@@ -11,7 +12,7 @@ TEST_CASE("ScenarioData::FromToml - push action creates StepPush") {
       "phase = \"wait\"\n"
       "duration = 1.5\n";
   const TOMLReader reader{MemoryViewReader{Toml.data(), Toml.size()}};
-  const ScenarioData data = ScenarioData::FromToml(reader);
+  const ScenarioData data = ScenarioData::FromToml(reader, GetPhaseLoaders());
   REQUIRE(data.sections.contains(U"intro"));
   REQUIRE(data.sections.at(U"intro").size() == 1);
   const auto& step = data.sections.at(U"intro")[0];
@@ -29,7 +30,7 @@ TEST_CASE("ScenarioData::FromToml - reset action creates StepReset") {
       "phase = \"wait\"\n"
       "duration = 2.0\n";
   const TOMLReader reader{MemoryViewReader{Toml.data(), Toml.size()}};
-  const ScenarioData data = ScenarioData::FromToml(reader);
+  const ScenarioData data = ScenarioData::FromToml(reader, GetPhaseLoaders());
   REQUIRE(data.sections.contains(U"intro"));
   REQUIRE(data.sections.at(U"intro").size() == 1);
   const auto& step = data.sections.at(U"intro")[0];
@@ -46,7 +47,7 @@ TEST_CASE("ScenarioData::FromToml - unknown phase name throws Error") {
       "action = \"push\"\n"
       "phase = \"nonexistent\"\n";
   const TOMLReader reader{MemoryViewReader{Toml.data(), Toml.size()}};
-  REQUIRE_THROWS_AS(ScenarioData::FromToml(reader), Error);
+  REQUIRE_THROWS_AS(ScenarioData::FromToml(reader, GetPhaseLoaders()), Error);
 }
 
 TEST_CASE("ScenarioData::FromToml - unknown action throws Error") {
@@ -55,7 +56,7 @@ TEST_CASE("ScenarioData::FromToml - unknown action throws Error") {
       "action = \"fly\"\n"
       "phase = \"wait\"\n";
   const TOMLReader reader{MemoryViewReader{Toml.data(), Toml.size()}};
-  REQUIRE_THROWS_AS(ScenarioData::FromToml(reader), Error);
+  REQUIRE_THROWS_AS(ScenarioData::FromToml(reader, GetPhaseLoaders()), Error);
 }
 
 TEST_CASE("ScenarioData::FromToml - missing duration throws Error") {
@@ -64,7 +65,7 @@ TEST_CASE("ScenarioData::FromToml - missing duration throws Error") {
       "action = \"push\"\n"
       "phase = \"wait\"\n";
   const TOMLReader reader{MemoryViewReader{Toml.data(), Toml.size()}};
-  REQUIRE_THROWS_AS(ScenarioData::FromToml(reader), Error);
+  REQUIRE_THROWS_AS(ScenarioData::FromToml(reader, GetPhaseLoaders()), Error);
 }
 
 TEST_CASE("ScenarioData::FromToml - missing param throws Error") {
@@ -73,7 +74,7 @@ TEST_CASE("ScenarioData::FromToml - missing param throws Error") {
       "action = \"push\"\n"
       "phase = \"scenario\"\n";
   const TOMLReader reader{MemoryViewReader{Toml.data(), Toml.size()}};
-  REQUIRE_THROWS_AS(ScenarioData::FromToml(reader), Error);
+  REQUIRE_THROWS_AS(ScenarioData::FromToml(reader, GetPhaseLoaders()), Error);
 }
 
 #endif
